Create Automatic water tap
An automatic tap using an Arduino, a servo motor, and an ultrasonic sensor can be designed to provide touchless operation. The ultrasonic sensor measures the distance to detect the presence of an object (such as a hand) near the tap, and the Arduino controls the servo motor to turn the tap on or off based on this information.